aibomgen-platform/scanner/requirements.txt
celery[sqlalchemy]
python-dotenv
docker
boto3
mysql-connector-python